move raw write and verify helpers into legacy_io
