+++ src/herder_daemon/writer_process/mod.rs
@@ -5,25 +5,20 @@
 
 use std::{
     fs::{File, OpenOptions},
-    io::{self, Read, Seek, Write},
+    io::{self, Read, Seek},
     os::unix::process::ExitStatusExt,
     process::{Command, Stdio},
     thread::JoinHandle,
 };
 
-use aligned_vec::avec_rt;
-use tracing::{debug, info, trace};
+use tracing::{debug, info};
 use tracing_unwrap::ResultExt;
 
 use self::{
-    utils::{CountRead, CountWrite, FileSourceReader, SyncDataFile},
     xplat::open_blockdev,
 };
-use crate::{compression::CompressionFormat, device, herder_api::write_verify::*};
+use crate::{ device, herder_api::write_verify::*, legacy_io::{SyncDataFile, VerifyOp, WriteOp}};
 
-#[cfg(test)]
-mod tests;
-mod utils;
 mod xplat;
 
 /// Maximum size we may allocate for each buffer.
@@ -168,153 +163,4 @@ fn run(
     .execute(tx)?;
 
     Ok(())
-}
-
-/// Wraps a bunch of parameters for a big complicated operation where we:
-///
-/// - decompress the input file
-/// - write to a disk
-/// - write stats down a pipe
-struct WriteOp<S: Read, D: Write> {
-    /// File to read from
-    file: S,
-    /// Disk to write to
-    disk: D,
-    /// Compression format to use
-    cf: CompressionFormat,
-    /// Buffer size to use when writing
-    buf_size: usize,
-    /// Block size of the disk
-    disk_block_size: usize,
-    /// How many writes of size [`Self::buf_size`] before we report back
-    checkpoint_period: usize,
-    /// How big the file reader's buffer should be
-    file_read_buf_size: usize,
-}
-
-impl<S: Read, D: Write> WriteOp<S, D> {
-    /// Execute the write operation. Returns total number of bytes written.
-    #[inline(always)]
-    fn execute(
-        &mut self,
-        mut tx: impl FnMut(WriteVerifyEvent),
-    ) -> Result<u64, LegacyWriteVerifyError> {
-        let mut file = FileSourceReader::new(self.cf, self.file_read_buf_size, &mut self.file);
-        let mut disk = CountWrite::new(&mut self.disk);
-        let mut buf = avec_rt![[self.disk_block_size] | 0u8; self.buf_size];
-
-        macro_rules! checkpoint {
-            () => {
-                tx(WriteVerifyEvent::TotalBytes {
-                    src: file.read_file_bytes(),
-                    dest: disk.count(),
-                });
-            };
-        }
-
-        loop {
-            for _ in 0..self.checkpoint_period {
-                // Try to fill up the block if we can.
-                let read_bytes = try_read_exact(&mut file, &mut buf)?;
-                if read_bytes == 0 {
-                    disk.flush()?;
-                    checkpoint!();
-                    return Ok(file.decompressed_bytes());
-                }
-
-                // Write the entire buffer, because we're doing direct writes.
-                // Even if we didn't fill the whole buffer, we are still writing the whole
-                // buffer.
-                let written_bytes = disk.write(&buf[..])?;
-                if written_bytes == 0 {
-                    checkpoint!();
-                    return Err(LegacyWriteVerifyError::EndOfOutput);
-                }
-            }
-            checkpoint!();
-        }
-    }
-}
-
-/// Like [`ReadExt::read_exact`], but if it can't fill the entire buffer, it
-/// does not error.
-#[inline(always)]
-fn try_read_exact(r: &mut impl Read, mut buf: &mut [u8]) -> std::io::Result<usize> {
-    // modified from rust stdlib file src/io/mod.rs
-
-    let orig_len = buf.len();
-    while !buf.is_empty() {
-        match r.read(buf) {
-            Ok(0) => break,
-            Ok(n) => {
-                buf = &mut buf[n..];
-            }
-            Err(e) => return Err(e),
-        }
-    }
-    Ok(orig_len - buf.len())
-}
-
-/// Wraps a bunch of parameters for a big complicated operation where we:
-///
-/// - decompress the input file
-/// - read from a disk
-/// - verify both sides are correct
-/// - write stats down a pipe
-struct VerifyOp<S: Read, D: Read> {
-    /// File to validate against
-    file: S,
-    /// Disk to validate
-    disk: D,
-    /// Compression format to use
-    cf: CompressionFormat,
-    /// Buffer size to use when writing
-    buf_size: usize,
-    /// Block size of the disk
-    disk_block_size: usize,
-    /// How many writes of size [`Self::buf_size`] before we report back
-    checkpoint_period: usize,
-    /// How big the file reader's buffer should be
-    file_read_buf_size: usize,
-}
-
-impl<S: Read, D: Read> VerifyOp<S, D> {
-    #[inline(always)]
-    fn execute(
-        &mut self,
-        mut tx: impl FnMut(WriteVerifyEvent),
-    ) -> Result<(), LegacyWriteVerifyError> {
-        let mut file = FileSourceReader::new(self.cf, self.file_read_buf_size, &mut self.file);
-        let mut disk = CountRead::new(&mut self.disk);
-
-        let mut file_buf = avec_rt![[self.disk_block_size] | 0u8; self.buf_size];
-        let mut disk_buf = avec_rt![[self.disk_block_size] | 0u8; self.buf_size];
-
-        macro_rules! checkpoint {
-            () => {
-                tx(WriteVerifyEvent::TotalBytes {
-                    src: file.read_file_bytes(),
-                    dest: disk.count(),
-                });
-            };
-        }
-
-        loop {
-            for _ in 0..self.checkpoint_period {
-                let file_read_bytes = try_read_exact(&mut file, &mut file_buf)?;
-                if file_read_bytes == 0 {
-                    checkpoint!();
-                    return Ok(());
-                }
-
-                try_read_exact(&mut disk, &mut disk_buf)?;
-
-                if file_buf[..file_read_bytes] != disk_buf[..file_read_bytes] {
-                    trace!(file_read_bytes, "verification failed");
-                    return Err(LegacyWriteVerifyError::VerificationFailed);
-                }
-            }
-            checkpoint!();
-        }
-    }
 }
