+++ src/legacy_io/mod.rs
@@ -1,3 +1,11 @@
 mod hash;
+#[cfg(test)]
+mod tests;
+mod utils;
+mod verify;
+mod write;
 
 pub use hash::do_file_hashing;
+pub use utils::SyncDataFile;
+pub use verify::VerifyOp;
+pub use write::WriteOp;
