+++ src/legacy_io/tests.rs
@@ -3,8 +3,9 @@ use pretty_assertions::assert_eq;
 use rand::{SeedableRng, rngs::SmallRng};
 use rstest::*;
 
+use crate::herder_api::write_verify::{LegacyWriteVerifyError, WriteVerifyEvent};
+
 use self::helpers::*;
-use super::*;
 
 #[test]
 fn write_op_works_with_emitted_events() {
@@ -77,6 +78,8 @@ fn write_misaligned_file_works(
 
 #[rstest]
 fn write_file_larger_than_disk(#[values(1032, 2000, 6000, 7000)] file_size: usize) {
+    use crate::herder_api::write_verify::LegacyWriteVerifyError;
+
     let test = WriteTest {
         file_size,
         buf_size: 512,
@@ -261,8 +264,11 @@ mod helpers {
 
     use rand::{Rng, SeedableRng, rngs::SmallRng};
 
-    use super::{CompressionFormat, VerifyOp, WriteOp};
-    use crate::herder_api::write_verify::*;
+    use crate::{
+        compression::CompressionFormat,
+        herder_api::write_verify::*,
+        legacy_io::{VerifyOp, WriteOp},
+    };
 
     /// Wraps an in-memory buffer and logs every single chunk of data written to
     /// it.
